リファクタリング: コンポーネントを整理
