--- ocaz-simple-browser/app/src/app/components/ImageThumbnail.js
@@ -1,23 +1,0 @@
-import styles from "./ImageThumbnail.module.css";
-
-export default function ImageThumbnail({ object, height, onClick = null }) {
-  const width = Math.floor(height * (object.image.width / object.image.height));
-  return (
-    <div className={styles.outer}>
-      <div
-        className={styles.inner}
-        style={{
-          width: `${width}px`,
-          height: `${height}px`,
-        }}
-      >
-        <img
-          className={styles.image}
-          src={`/api/forwarder/object/head10mbSha1/${object.head10mbSha1}`}
-          loading="lazy"
-          onClick={onClick}
-        />
-      </div>
-    </div>
-  );
-}
