+++ ocaz-simple-browser/app/src/app/components/Thumbnail.module.css
@@ -6,14 +6,23 @@
   position: relative;
 }
 
-.image {
+.inner img,
+.inner video {
   position: absolute;
   top: 0;
   bottom: 0;
   left: 0;
   right: 0;
-  width: auto;
+  width: 100%;
   height: 100%;
   object-fit: contain;
   cursor: pointer;
+}
+
+.videoIcon {
+  position: absolute;
+  top: 3px;
+  left: 6px;
+  pointer-events: none;
+  opacity: 0.8;
 }
