--- ocaz-simple-browser/app/src/app/components/ThumbnailContainer.js
@@ -1,13 +1,0 @@
-export default function ThumbnailContainer({ children }) {
-  return (
-    <div
-      style={{
-        display: "flex",
-        flexWrap: "wrap",
-        justifyContent: "space-between",
-      }}
-    >
-      {children}
-    </div>
-  );
-}
