--- ocaz-simple-browser/app/src/app/components/VideoThumbnail.module.css
@@ -1,27 +1,0 @@
-.outer {
-  border: 2px solid #333333;
-}
-
-.inner {
-  position: relative;
-}
-
-.video {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  width: 100%;
-  height: auto;
-  object-fit: contain;
-  cursor: pointer;
-}
-
-.videoIcon {
-  position: absolute;
-  top: 3px;
-  left: 6px;
-  pointer-events: none;
-  opacity: 0.8;
-}
